+++ Assets/02. Scipts/InGame/Monster/MonsterData/MonsterData.cs
@@ -4,15 +4,15 @@ using UnityEngine;
 public class MonsterData : ScriptableObject
 {
     [Header("체력")]
-    [SerializeField] private float _maxHealth;
+    [SerializeField] private int _maxHealth;
 
     [Header("공격력")]
-    [SerializeField] private float _attackPower;
+    [SerializeField] private int _attackPower;
 
     [Header("이동 속도")]
     [SerializeField] private float _moveSpeed;
 
-    public float MaxHealth => _maxHealth;
-    public float AttackPower => _attackPower;
+    public int MaxHealth => _maxHealth;
+    public int AttackPower => _attackPower;
     public float MoveSpeed => _moveSpeed;
 }
